Marital Status, language Proficiency and Some Refactoring

+++ backend/src/main/java/com/inner/satisfaction/backend/authentication/token/AuthenticationToken.java
@@ -25,9 +25,10 @@ public class AuthenticationToken extends BaseEntity implements Authentication {
   private String user;
   private String token;
   // Can login from any one company
-  private long companyId;
+  private int companyId;
   private boolean isActive;
   private Timestamp expiry;
+
   @Convert(converter = JpaConverterJson.class)
   private List<String> roles;
